Move out some code
