+++ okhttp/src/jvmMain/kotlin/okhttp3/FormBody.kt
@@ -17,10 +17,10 @@ package okhttp3
 
 import java.io.IOException
 import java.nio.charset.Charset
-import okhttp3.HttpUrl.Companion.FORM_ENCODE_SET
-import okhttp3.HttpUrl.Companion.canonicalize
-import okhttp3.HttpUrl.Companion.percentDecode
 import okhttp3.MediaType.Companion.toMediaType
+import okhttp3.internal.JvmHttpUrl.FORM_ENCODE_SET
+import okhttp3.internal.JvmHttpUrl.canonicalize
+import okhttp3.internal.JvmHttpUrl.percentDecode
 import okhttp3.internal.toImmutableList
 import okio.Buffer
 import okio.BufferedSink
